consolidate store items
