--- src/store/user.js
@@ -1,68 +1,0 @@
-import axios from "axios";
-
-/**
- * ACTION TYPES
- */
-const GET_USER = "GET_USER";
-const REMOVE_USER = "REMOVE_USER";
-
-/**
- * INITIAL STATE
- */
-const defaultUser = {};
-
-/**
- * ACTION CREATORS
- */
-const getUser = user => ({ type: GET_USER, user });
-const removeUser = () => ({ type: REMOVE_USER });
-
-/**
- * THUNK CREATORS
- */
-export const me = () => async dispatch => {
-  try {
-    //going to need some async here
-    dispatch(getUser(res.data || defaultUser));
-  } catch (err) {
-    console.error(err);
-  }
-};
-
-export const auth = (email, password, method) => async dispatch => {
-  let res;
-  try {
-    //going to need some async here
-  } catch (authError) {
-    return dispatch(getUser({ error: authError }));
-  }
-
-  try {
-    dispatch(getUser(res.data));
-  } catch (dispatchOrHistoryErr) {
-    console.error(dispatchOrHistoryErr); //this is just a fancy way to label the error
-  }
-};
-
-export const logout = () => async dispatch => {
-  try {
-    //going to need some async here
-    dispatch(removeUser());
-  } catch (err) {
-    console.error(err);
-  }
-};
-
-/**
- * REDUCER
- */
-export default function(state = defaultUser, action) {
-  switch (action.type) {
-    case GET_USER:
-      return action.user;
-    case REMOVE_USER:
-      return defaultUser;
-    default:
-      return state;
-  }
-}
